Add files via upload
A critical task that must be done every week is ensuring all students apart of the PCDC youth afterschool program of 90+ highschoolers meet the minimum attendance requirements of 3 days a week. There are circumstances where youth may need to be excused or sports season where they are only required to show up once or twice a week. For special circumstances youth must be present for at least 2 hours a week for one day or two days 1 hour each.

Normally all of this is done manually and takes the average person around 3 hours after double checking. What I created was something that could automate the process and simplify it so all the user would need to do is double check and that is all.

What I built
PART 1
* Finds relevant part of dataset from cityspan csv export
*Iterates through all the rows with youth names, attendance and on which days and displays it.

PART 2
After dataset is confirmed to show all the students in the list and is accurate
part 2 grabs all the emails in csv form extracted from the actual excel sheet for data entry
*Merges all youth name with their emails for later use
*Adds another data set with student names, amount of SY Master Hours attended and days
*Shows a list of all 0 and 1 absences and the hours to verify if they met requirements for special circumstances and
 shows emails for the email to be sent out to youth
to ensure they make up said hours.